requirements.txt
investpy
nest-asyncio
ib-insync
# =============================================================================
# ALPHA PLATFORM - Python Dependencies
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pgvector>=0.2.0

# -----------------------------------------------------------------------------
# Kafka / Message Queue
# -----------------------------------------------------------------------------
kafka-python>=2.0.2
# Alternative: confluent-kafka>=2.3.0

# -----------------------------------------------------------------------------
# Data & Analysis
# -----------------------------------------------------------------------------
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
scipy>=1.11.0
scikit-learn>=1.3.0

# -----------------------------------------------------------------------------
# Web & API
# -----------------------------------------------------------------------------
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0

# -----------------------------------------------------------------------------
# LLM / AI
# -----------------------------------------------------------------------------
openai>=1.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# -----------------------------------------------------------------------------
# Dashboard
# -----------------------------------------------------------------------------
streamlit>=1.29.0
plotly>=5.18.0
altair>=5.2.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.66.0
tenacity>=8.2.0
cachetools>=5.3.0

# -----------------------------------------------------------------------------
# Logging & Monitoring
# -----------------------------------------------------------------------------
structlog>=23.2.0
rich>=13.7.0

# -----------------------------------------------------------------------------
# Testing (optional, for development)
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# -----------------------------------------------------------------------------
# Development Tools (optional)
# -----------------------------------------------------------------------------
black>=23.12.0
isort>=5.13.0
mypy>=1.7.0
